Add @table only when using non-stanard naming
